Simplification of prospect.sources
Moves the SSPBasis and FastStepBasis to galaxy_basis module.
Removes the ssp_basis, dust_basis, and boneyard modules.
Removes the get_spectrum method from SSPBasis and subclasses; this is handled by SpecModel
Remove various multicomponent hacks.
