+++ prospect/sources/__init__.py
@@ -1,12 +1,4 @@
 from .galaxy_basis import *
-from .ssp_basis import *
-from .star_basis import *
-from .dust_basis import *
-from .boneyard import StepSFHBasis
 
-__all__ = ["to_cgs",
-           "CSPSpecBasis", "MultiComponentCSPBasis",
-           "FastSSPBasis", "SSPBasis",
-           "FastStepBasis", "StepSFHBasis",
-           "StarBasis", "BigStarBasis",
-           "BlackBodyDustBasis"]
+__all__ = ["CSPSpecBasis", "SSPBasis",
+           "FastStepBasis"]
